Only hook CreateFileW This changes the loader dll to only hook CreateFileW, instead of also hooking ReadFile and CloseHandle.
This was done to ensure that all checks for file size and similar would
be consistent.

This also simplifies a lot, because now we are not responsible for
closing the data.win handles, we just replace the path with our own
data.win instead, and the runner handles closing them.